refactor(lex-lsp): split server.rs into cohesive submodules
